Fixes and improvements to threshold finding function
this commit fixes an issue where the threshold finding algorithm fails
to find the correct threshold for traces where the noise standard
deviation is smaller than the mean value of the beginning of the trace.

Additionally the performance of the function is improved by using numpy.searchsorted instead of boolean array indexing. (timeit
measurements show a performance improvement of about factor of 6.)